// src/shared/types/hero.types.ts
export interface Origin {
  name: string;
  url: string;
}

export interface Location {
  name: string;
  url: string;
}
export interface Hero {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: Origin;
  location: Location;
  image: string;
  episode: string[];
  url: string;
  created: string;
}